// DANCEfinder/src/app/models/IErrorTextMessages.ts
export interface IRegistrationErrorTextMessages {
  name: string;
  surname: string;
  email: string;
  password: string;
  phone: string;
  sex: string;
  group: string;
}

export interface IGroupErrorTextMessages {
  instructor: string;
  noInstructor: string;
  classroom: string;
  noClassroom: string;
  startDate: string;
}

export interface ISchoolErrorTextMessages {
  schoolName: string;
  postalCode: string;
  city: string;
  street: string;
  website: string;
  phoneNumber: string;
  email: string;
  about: string;
}

export interface IClassroomErrorTextMessages {
  number: string;
}

export interface ICourseErrorTextMessages {
  name: string;
  level: string;
  price: string;
  dance: string;
}

export interface IInstructorErrorTextMessages {
  instructor: string;
}

export interface IStudentErrorTextMessages {
  student: string;
}

export interface IAccountErrorTextMessages {
  name: string;
  surname: string;
  phone: string;
  sex: string;
}

export interface IPaymentTextMessages {
  accountNumber: string;
}